+++ src/App.tsx
@@ -13,29 +13,33 @@ import * as _ from "lodash";
 import { H1, H2, H3 } from "./components/heading";
 import { getData, getReportStructure } from "./lib/get-data";
 import { Histogram } from "./components/histogram";
-import { FilterSettingsForm } from "./components/filter-settings-form";
-import { DataSummaryTable } from "./components/data-summary-table";
+import { FilterSettingsForm } from "./components/app/filter-settings-form";
+import { DataSummaryTable } from "./components/app/data-summary-table";
 import { BarPlot } from "./components/barplot";
 import { ScatterPlot } from "./components/scatterplot";
 import { createMemo } from "solid-js";
-import { SampleFilterForm } from "./components/sample-filter-form";
+import { SampleFilterForm } from "./components/app/sample-filter-form";
 import { filterData, calculateQcPassCells, getPassingCellIndices } from "./lib/data-filters";
 import { transformSampleMetadata } from "./lib/sample-utils";
 import { hasSpatialCoordinates } from "./lib/plots";
+import { createSettingsForm, SettingsFormProvider } from "./components/app/settings-form";
+import { GlobalVisualizationSettings } from "./components/app/global-visualization-settings";
 
 
 const App: Component = () => {
   const [reportStructure, setReportStructure] = createSignal<ReportStructure>({categories: []});
   const [data, setData] = createSignal<RawData>();
-  const [selectedSamples, setSelectedSamples] = createSignal<string[]>([]);
-  const [globalGroupBy, setGlobalGroupBy] = createSignal<string>("sample_id");
 
-  // Add a new signal to track if global grouping is enabled
-  const [isGlobalGroupingEnabled, setIsGlobalGroupingEnabled] = createSignal(true);
   const [filtersApplied, setFiltersApplied] = createSignal(false);
 
   // Add a new state to store the applied filter settings
   const [appliedFilterSettings, setAppliedFilterSettings] = createStore<Settings>({});
+
+  const form = createSettingsForm();
+
+  const globalVisualization = form.useStore(state => state.values.globalVisualization);
+  const selectedSamples = form.useStore(state => state.values.sampleSelection.selectedSamples);
+
 
   // read data in memory
   createEffect(async () => {
@@ -220,287 +224,256 @@ const App: Component = () => {
 
   // page layout
   return (
-    <div class="container mx-a space-y-2">
-      <H1>OpenPipelines Ingestion QC Report</H1>
-      <SampleFilterForm
-        sampleIds={data()?.sample_summary_stats.columns.find(c => c.name === "sample_id")?.categories || []}
-        selectedSamples={selectedSamples()}
-        onChange={setSelectedSamples}
-        sampleMetadata={transformSampleMetadata(data())}
-      />
-      <div class="mb-4 p-4 bg-gray-50 rounded-md border">
-        <h3 class="text-lg font-medium mb-2">Global Visualization Settings</h3>
-        <div class="flex items-center gap-2 mb-2">
-          <div class="flex items-center">
-            <input 
-              type="checkbox" 
-              id="enable-global-grouping"
-              checked={isGlobalGroupingEnabled()} 
-              onChange={(e) => setIsGlobalGroupingEnabled(e.target.checked)}
-              class="mr-2 h-4 w-4"
-            />
-            <label for="enable-global-grouping" class="w-auto text-sm font-medium">
-              Enable global grouping
-            </label>
-          </div>
-        </div>
-        <div class="flex items-center gap-2">
-          <label class="w-24">Group By:</label>
-          <select 
-            class="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background"
-            value={globalGroupBy()}
-            onChange={(e) => setGlobalGroupBy(e.target.value)}
-            disabled={!isGlobalGroupingEnabled()}
-          >
-            <For each={getCategoricalColumns()}>
-              {(column) => (
-                <option value={column}>{column}</option>
-              )}
-            </For>
-          </select>
-        </div>
-      </div>
-      <For each={reportStructure().categories}>
-        {(category) => (
-          <Show when={(settings[category.key] || []).length > 0}>
-            <H2>{category.name}</H2>
-            
-            {/* Add descriptive text based on category */}
-            <div class="mb-4 text-gray-700">
-              {category.key === "sample_summary_stats" && (
-                <p>
-                  These metrics provide sample-level quality control measures. Each plot represents data 
-                  aggregated at the sample level and is always grouped by sample ID.
-                </p>
-              )}
-              
-              {category.key === "metrics_cellranger_stats" && (
-                <p>
-                  These metrics are provided by CellRanger and show sample-level sequencing statistics. 
-                  All CellRanger metrics are always grouped by sample ID regardless of global visualization settings.
-                </p>
-              )}
-              
-              {category.key === "cell_rna_stats" && (
-                <p>
-                  These metrics provide cell-level quality control measures. You can use the Global Visualization 
-                  Settings above to group these metrics by different categorical variables (e.g., sample_id, 
-                  predicted cell type). These groupings allow you to identify quality differences across 
-                  biological or technical groups.
-                </p>
-              )}
-            </div>
-            
-            <div class="grid grid-cols-1 gap-4">
-              <For each={settings[category.key]}>
-                {(_, i) => {
-                  // Extract the groupBy logic into a reactive memo
-                  const currentFilterGroupBy = createMemo(() => {
-                    if (category.key === "metrics_cellranger_stats") {
-                      return "sample_id"; // CellRanger metrics always use sample_id
-                    } else if (isGlobalGroupingEnabled()) {
-                      return globalGroupBy(); // Use global setting when enabled
-                    } else {
-                      return settings[category.key][i()].groupBy || "sample_id"; // Use plot's own setting or default
-                    }
-                  });
-                  
-                  const [isExpanded, setIsExpanded] = createSignal(true); // Start expanded
-                  
-                  return (
-                    <div>
-                      <div class="flex justify-between items-center mb-2">
-                        <H3>{settings[category.key][i()].label}</H3>
-                        
-                        {/* Add the visualization toggle in the top-right corner */}
-                        <Show when={category.key === "cell_rna_stats" && 
-                                   settings[category.key][i()].type === "histogram" && 
-                                   hasSpatialCoordinates(data()?.cell_rna_stats)}>
-                          <div 
-                            class="relative rounded-full bg-gray-200 shadow-sm overflow-hidden"
-                            style={{ height: "32px", width: "180px" }}
-                          >
-                            <div 
-                              class="absolute bg-white rounded-full shadow transition-transform duration-200"
-                              style={{
-                                width: "calc(50% - 4px)",
-                                height: "calc(100% - 4px)",
-                                top: "2px",
-                                left: "2px",
-                                transform: settings[category.key][i()].visualizationType !== 'spatial' 
-                                  ? 'translateX(0)' 
-                                  : 'translateX(calc(100% + 4px))'
-                              }}
-                            />
-                            
-                            <div class="absolute inset-0 flex w-full h-full">
-                              <div 
-                                class="flex items-center justify-center w-1/2 cursor-pointer"
-                                onClick={() => setSettings(category.key, i(), { ...settings[category.key][i()], visualizationType: 'histogram' })}
-                              >
-                                <span 
-                                  class={`text-sm font-medium transition-colors duration-200 ${
-                                    settings[category.key][i()].visualizationType !== 'spatial' ? 'text-gray-800' : 'text-gray-500'
-                                  }`}
-                                >
-                                  Histogram
-                                </span>
-                              </div>
-                              
-                              <div 
-                                class="flex items-center justify-center w-1/2 cursor-pointer"
-                                onClick={() => setSettings(category.key, i(), { ...settings[category.key][i()], visualizationType: 'spatial' })}
-                              >
-                                <span 
-                                  class={`text-sm font-medium transition-colors duration-200 ${
-                                    settings[category.key][i()].visualizationType === 'spatial' ? 'text-gray-800' : 'text-gray-500'
-                                  }`}
-                                >
-                                  Spatial
-                                </span>
-                              </div>
-                            </div>
-                          </div>
-                        </Show>
-                      </div>
-                      
-                      <Show when={settings[category.key][i()].description}>
-                        <p class="text-gray-600 text-sm mb-2">{settings[category.key][i()].description}</p>
-                      </Show>
-
-                      <button 
-                        onClick={() => setIsExpanded(!isExpanded())}
-                        class="w-full px-4 py-2 text-left text-sm font-medium text-gray-700 bg-gray-100 hover:bg-gray-200 rounded-md flex justify-between items-center mb-2"
-                      >
-                        <span>Plot Visibility</span>
-                        <span class="transition-transform duration-200" classList={{ "rotate-180": !isExpanded() }}>
-                          ▼
-                        </span>
-                      </button>
-                      
-                      {isExpanded() && (
-                        <div class="flex flex-col space-y-2">
-                          <Switch>
-                            <Match when={!data()}>
-                              <div>Loading...</div>
-                            </Match>
-                            <Match when={settings[category.key][i()].type === "bar"}>
-                              <BarPlot
-                                data={(filtersApplied() ? fullyFilteredData() : filteredData())![category.key]}
-                                filterSettings={{
-                                  ...settings[category.key][i()],
-                                  groupBy: currentFilterGroupBy()
-                                }}
-                              />
-                            </Match>
-                            <Match when={settings[category.key][i()].type === "histogram" && 
-                                         (settings[category.key][i()].visualizationType === "histogram" || !settings[category.key][i()].visualizationType)}>
-                              <Histogram
-                                data={(filtersApplied() ? fullyFilteredData() : filteredData())![category.key]}
-                                filterSettings={{
-                                  ...settings[category.key][i()],
-                                  groupBy: currentFilterGroupBy()
-                                }}
-                                additionalAxes={category.additionalAxes}
-                              />
-                            </Match>
-                            <Match when={settings[category.key][i()].type === "histogram" && 
-                                         settings[category.key][i()].visualizationType === "spatial"}>
-                              <ScatterPlot
-                                data={(filtersApplied() ? fullyFilteredData() : filteredData())![category.key]}
-                                filterSettings={{
-                                  ...settings[category.key][i()],
-                                  groupBy: currentFilterGroupBy()
-                                }}
-                                additionalAxes={category.additionalAxes}
-                                colorFieldName={settings[category.key][i()].field}
-                              />
-                            </Match>
-                          </Switch>
-                          <FilterSettingsForm
-                            filterSettings={settings[category.key][i()]}
-                            updateFilterSettings={(fn) =>
-                              setSettings(category.key, i(), produce(fn))
-                            }
-                            data={(filtersApplied() ? fullyFilteredData() : filteredData())![category.key]}
-                            globalGroupBy={category.key === "metrics_cellranger_stats" ? undefined : (isGlobalGroupingEnabled() ? globalGroupBy() : undefined)}
-                            forceGroupBy={category.key === "metrics_cellranger_stats" ? "sample_id" : undefined}
-                            isGlobalGroupingEnabled={isGlobalGroupingEnabled()}
-                            category={category.key} // Pass the category key
-                          />
-                        </div>
-                      )}
-                    </div>
-                  );
-                }}
-              </For>
-            </div>
-          </Show>
-        )}
-      </For>
-      <div>
-        <H2>Results</H2>
-        <Show when={data()} fallback={<p># Cells before filtering: ...</p>}>
-          <p># Cells before filtering: {data()!.cell_rna_stats.num_rows}</p>
-        </Show>
-        <Show when={data()} fallback={<p># Cells after filtering: ...</p>}>
-          <p># Cells after filtering: {qcPass()}</p>
-          <div class="mt-4 flex gap-2">
-            <button 
-              onClick={() => {
-                // Take a snapshot of the current settings and save it as the applied settings
-                setAppliedFilterSettings(JSON.parse(JSON.stringify(settings)));
-                setFiltersApplied(true);
-              }}
-              class="px-4 py-2 bg-blue-600 text-white rounded-md hover:bg-blue-700 transition-colors"
-            >
-              Apply Filters to Plots
-            </button>
-            
-            <button 
-              onClick={() => {
-                setFiltersApplied(false);
-                // No need to clear applied settings - they'll be ignored when filtersApplied is false
-              }}
-              class="px-4 py-2 bg-gray-600 text-white rounded-md hover:bg-gray-700 transition-colors"
-            >
-              Reset to Default View
-            </button>
-            
-            <button 
-              onClick={exportFiltersAsYaml}
-              class="px-4 py-2 bg-green-600 text-white rounded-md hover:bg-green-700 transition-colors"
-            >
-              Export Filters as YAML
-            </button>
-            
-            {filtersApplied() && (
-              <p class="text-sm text-green-600 flex items-center">
-                ✓ Filters applied - Plots show only cells that pass all thresholds
-              </p>
-            )}
-          </div>
-        </Show>
-      </div>
-      <div>
-        <H2>Overview of loaded data</H2>
-        <p>
-          This overview is meant to give a quick glance at the data that has
-          been loaded.
-        </p>
+    <SettingsFormProvider form={form}>
+      <div class="container mx-a space-y-2">
+        <H1>OpenPipelines Ingestion QC Report</H1>
+        <SampleFilterForm
+          sampleIds={data()?.sample_summary_stats.columns.find(c => c.name === "sample_id")?.categories || []}
+          sampleMetadata={transformSampleMetadata(data())}
+        />
+        <GlobalVisualizationSettings getCategoricalColumns={getCategoricalColumns} />
         <For each={reportStructure().categories}>
           {(category) => (
-            <div>
-              <H3>{category.name}</H3>
-              <Show when={data()}>
-                <DataSummaryTable data={(filtersApplied() ? fullyFilteredData() : filteredData())![category.key]} />
-              </Show>
-            </div>
+            <Show when={(settings[category.key] || []).length > 0}>
+              <H2>{category.name}</H2>
+              
+              {/* Add descriptive text based on category */}
+              <div class="mb-4 text-gray-700">
+                {category.key === "sample_summary_stats" && (
+                  <p>
+                    These metrics provide sample-level quality control measures. Each plot represents data 
+                    aggregated at the sample level and is always grouped by sample ID.
+                  </p>
+                )}
+                
+                {category.key === "metrics_cellranger_stats" && (
+                  <p>
+                    These metrics are provided by CellRanger and show sample-level sequencing statistics. 
+                    All CellRanger metrics are always grouped by sample ID regardless of global visualization settings.
+                  </p>
+                )}
+                
+                {category.key === "cell_rna_stats" && (
+                  <p>
+                    These metrics provide cell-level quality control measures. You can use the Global Visualization 
+                    Settings above to group these metrics by different categorical variables (e.g., sample_id, 
+                    predicted cell type). These groupings allow you to identify quality differences across 
+                    biological or technical groups.
+                  </p>
+                )}
+              </div>
+              
+              <div class="grid grid-cols-1 gap-4">
+                <For each={settings[category.key]}>
+                  {(_, i) => {
+                    // Extract the groupBy logic into a reactive memo
+                    const currentFilterGroupBy = createMemo(() => {
+                      if (category.key === "metrics_cellranger_stats") {
+                        return "sample_id"; // CellRanger metrics always use sample_id
+                      } else if (globalVisualization().groupingEnabled) {
+                        return globalVisualization().groupBy; // Use global setting when enabled
+                      } else {
+                        return settings[category.key][i()].groupBy || "sample_id"; // Use plot's own setting or default
+                      }
+                    });
+                    
+                    const [isExpanded, setIsExpanded] = createSignal(true); // Start expanded
+                    
+                    return (
+                      <div>
+                        <div class="flex justify-between items-center mb-2">
+                          <H3>{settings[category.key][i()].label}</H3>
+                          
+                          {/* Add the visualization toggle in the top-right corner */}
+                          <Show when={category.key === "cell_rna_stats" && 
+                                    settings[category.key][i()].type === "histogram" && 
+                                    hasSpatialCoordinates(data()?.cell_rna_stats)}>
+                            <div 
+                              class="relative rounded-full bg-gray-200 shadow-sm overflow-hidden"
+                              style={{ height: "32px", width: "180px" }}
+                            >
+                              <div 
+                                class="absolute bg-white rounded-full shadow transition-transform duration-200"
+                                style={{
+                                  width: "calc(50% - 4px)",
+                                  height: "calc(100% - 4px)",
+                                  top: "2px",
+                                  left: "2px",
+                                  transform: settings[category.key][i()].visualizationType !== 'spatial' 
+                                    ? 'translateX(0)' 
+                                    : 'translateX(calc(100% + 4px))'
+                                }}
+                              />
+                              
+                              <div class="absolute inset-0 flex w-full h-full">
+                                <div 
+                                  class="flex items-center justify-center w-1/2 cursor-pointer"
+                                  onClick={() => setSettings(category.key, i(), { ...settings[category.key][i()], visualizationType: 'histogram' })}
+                                >
+                                  <span 
+                                    class={`text-sm font-medium transition-colors duration-200 ${
+                                      settings[category.key][i()].visualizationType !== 'spatial' ? 'text-gray-800' : 'text-gray-500'
+                                    }`}
+                                  >
+                                    Histogram
+                                  </span>
+                                </div>
+                                
+                                <div 
+                                  class="flex items-center justify-center w-1/2 cursor-pointer"
+                                  onClick={() => setSettings(category.key, i(), { ...settings[category.key][i()], visualizationType: 'spatial' })}
+                                >
+                                  <span 
+                                    class={`text-sm font-medium transition-colors duration-200 ${
+                                      settings[category.key][i()].visualizationType === 'spatial' ? 'text-gray-800' : 'text-gray-500'
+                                    }`}
+                                  >
+                                    Spatial
+                                  </span>
+                                </div>
+                              </div>
+                            </div>
+                          </Show>
+                        </div>
+                        
+                        <Show when={settings[category.key][i()].description}>
+                          <p class="text-gray-600 text-sm mb-2">{settings[category.key][i()].description}</p>
+                        </Show>
+
+                        <button 
+                          onClick={() => setIsExpanded(!isExpanded())}
+                          class="w-full px-4 py-2 text-left text-sm font-medium text-gray-700 bg-gray-100 hover:bg-gray-200 rounded-md flex justify-between items-center mb-2"
+                        >
+                          <span>Plot Visibility</span>
+                          <span class="transition-transform duration-200" classList={{ "rotate-180": !isExpanded() }}>
+                            ▼
+                          </span>
+                        </button>
+                        
+                        {isExpanded() && (
+                          <div class="flex flex-col space-y-2">
+                            <Switch>
+                              <Match when={!data()}>
+                                <div>Loading...</div>
+                              </Match>
+                              <Match when={settings[category.key][i()].type === "bar"}>
+                                <BarPlot
+                                  data={(filtersApplied() ? fullyFilteredData() : filteredData())![category.key]}
+                                  filterSettings={{
+                                    ...settings[category.key][i()],
+                                    groupBy: currentFilterGroupBy()
+                                  }}
+                                />
+                              </Match>
+                              <Match when={settings[category.key][i()].type === "histogram" && 
+                                          (settings[category.key][i()].visualizationType === "histogram" || !settings[category.key][i()].visualizationType)}>
+                                <Histogram
+                                  data={(filtersApplied() ? fullyFilteredData() : filteredData())![category.key]}
+                                  filterSettings={{
+                                    ...settings[category.key][i()],
+                                    groupBy: currentFilterGroupBy()
+                                  }}
+                                  additionalAxes={category.additionalAxes}
+                                />
+                              </Match>
+                              <Match when={settings[category.key][i()].type === "histogram" && 
+                                          settings[category.key][i()].visualizationType === "spatial"}>
+                                <ScatterPlot
+                                  data={(filtersApplied() ? fullyFilteredData() : filteredData())![category.key]}
+                                  filterSettings={{
+                                    ...settings[category.key][i()],
+                                    groupBy: currentFilterGroupBy()
+                                  }}
+                                  additionalAxes={category.additionalAxes}
+                                  colorFieldName={settings[category.key][i()].field}
+                                />
+                              </Match>
+                            </Switch>
+                            <FilterSettingsForm
+                              filterSettings={settings[category.key][i()]}
+                              updateFilterSettings={(fn) =>
+                                setSettings(category.key, i(), produce(fn))
+                              }
+                              data={(filtersApplied() ? fullyFilteredData() : filteredData())![category.key]}
+                              globalGroupBy={category.key === "metrics_cellranger_stats" ? undefined : (globalVisualization().groupingEnabled ? globalVisualization().groupBy : undefined)}
+                              forceGroupBy={category.key === "metrics_cellranger_stats" ? "sample_id" : undefined}
+                              isGlobalGroupingEnabled={globalVisualization().groupingEnabled}
+                              category={category.key} // Pass the category key
+                            />
+                          </div>
+                        )}
+                      </div>
+                    );
+                  }}
+                </For>
+              </div>
+            </Show>
           )}
         </For>
+        <div>
+          <H2>Results</H2>
+          <Show when={data()} fallback={<p># Cells before filtering: ...</p>}>
+            <p># Cells before filtering: {data()!.cell_rna_stats.num_rows}</p>
+          </Show>
+          <Show when={data()} fallback={<p># Cells after filtering: ...</p>}>
+            <p># Cells after filtering: {qcPass()}</p>
+            <div class="mt-4 flex gap-2">
+              <button 
+                onClick={() => {
+                  // Take a snapshot of the current settings and save it as the applied settings
+                  setAppliedFilterSettings(JSON.parse(JSON.stringify(settings)));
+                  setFiltersApplied(true);
+                }}
+                class="px-4 py-2 bg-blue-600 text-white rounded-md hover:bg-blue-700 transition-colors"
+              >
+                Apply Filters to Plots
+              </button>
+              
+              <button 
+                onClick={() => {
+                  setFiltersApplied(false);
+                  // No need to clear applied settings - they'll be ignored when filtersApplied is false
+                }}
+                class="px-4 py-2 bg-gray-600 text-white rounded-md hover:bg-gray-700 transition-colors"
+              >
+                Reset to Default View
+              </button>
+              
+              <button 
+                onClick={exportFiltersAsYaml}
+                class="px-4 py-2 bg-green-600 text-white rounded-md hover:bg-green-700 transition-colors"
+              >
+                Export Filters as YAML
+              </button>
+              
+              {filtersApplied() && (
+                <p class="text-sm text-green-600 flex items-center">
+                  ✓ Filters applied - Plots show only cells that pass all thresholds
+                </p>
+              )}
+            </div>
+          </Show>
+        </div>
+        <div>
+          <H2>Overview of loaded data</H2>
+          <p>
+            This overview is meant to give a quick glance at the data that has
+            been loaded.
+          </p>
+          <For each={reportStructure().categories}>
+            {(category) => (
+              <div>
+                <H3>{category.name}</H3>
+                <Show when={data()}>
+                  <DataSummaryTable data={(filtersApplied() ? fullyFilteredData() : filteredData())![category.key]} />
+                </Show>
+              </div>
+            )}
+          </For>
+        </div>
+        <div class="h-64" />
       </div>
-      <div class="h-64" />
-    </div>
+    </SettingsFormProvider>
   );
 };
