+++ src/components/app/settings-form.tsx
@@ -1,0 +1,55 @@
+import { createForm } from "@tanstack/solid-form";
+import { createContext, ParentComponent, useContext } from "solid-js";
+
+export type SettingsState = {
+  sampleSelection: {
+    selectedSamples: string[];
+  }
+  globalVisualization: {
+    groupingEnabled: boolean;
+    groupBy: string;
+  }
+}
+
+export const defaultSettings: SettingsState = {
+  sampleSelection: {
+    selectedSamples: []
+  },
+  globalVisualization: {
+    groupingEnabled: true,
+    groupBy: "sample_id",
+  }
+}
+
+export function createSettingsForm () {
+  const form = createForm(() => ({
+    defaultValues: defaultSettings,
+    listeners: {
+      onChange: ({formApi}) => {
+        console.log("Settings changed:", formApi.state.values);
+      }
+    }
+  }))
+
+  return form
+}
+
+export type SettingsFormReturnType = ReturnType<typeof createSettingsForm>;
+
+// we create a context to provide the form to many components
+const SettingsFormContext = createContext<SettingsFormReturnType>();
+
+export const SettingsFormProvider: ParentComponent<{form: SettingsFormReturnType}> = (props) => {
+  return (
+    <SettingsFormContext.Provider value={props.form}>
+      {props.children}
+    </SettingsFormContext.Provider>
+  );
+}
+export const useSettingsForm = () => {
+  const form = useContext(SettingsFormContext);
+  if (!form) {
+    throw new Error("useSettingsForm must be used within a SettingsFormProvider");
+  }
+  return form;
+}
